+++ src/components/hoc/center.tsx
@@ -1,10 +1,12 @@
+import { memo } from "react";
+
 type Props = {
     children: any
 }
-export const Center = ({ children }: Props) => {
+export const Center = memo(({ children }: Props) => {
     return (
         <div className="center">
             {children}
         </div>
     );
-}
+})
