+++ src/components/hoc/right.tsx
@@ -1,10 +1,12 @@
+import { memo } from "react";
+
 type Props = {
     children: any
 }
-export const Right = ({ children }: Props) => {
+export const Right = memo(({ children }: Props) => {
     return (
         <div className="right">
             {children}
         </div>
     );
-}
+})
