+++ src/components/hoc/rightPanel.tsx
@@ -1,10 +1,12 @@
+import { memo } from "react";
+
 type Props = {
     children: any
 }
-export const RightPanel = ({ children }: Props) => {
+export const RightPanel = memo(({ children }: Props) => {
     return (
         <div className="rightpanel">
             {children}
         </div>
     );
-}
+})
