+++ src/components/ui/meny-time use/loader.tsx
@@ -21,7 +21,7 @@ export const MainLoader = () => {
     }, [points])
     return (
         <>
-            <div style={{ position: 'fixed', backgroundColor: '#242424', inset: '0', zIndex: '10' }} >
+            <div style={{ position: 'fixed', backgroundColor: '#24242490', inset: '0', zIndex: '10', animation: 'none', backdropFilter: 'blur(10px)' }} className="shadow">
                 <div style={{ ...positioncenterbyabsolute, top: '37%' }} className="mainloader">
                     <div className="dropswadow">
                         <EmpieLogo size={8.6} />
