+++ src/pages/direction/Meetings.tsx
@@ -1,23 +1,26 @@
 import { useContext, useState } from "react";
-import { ModalDirectionChildren } from "../../childrens/modalDirection";
+import { ModalDirectionChildren } from "../../childrens/other/modalDirection";
 import { LeftPanel } from "../../components/hoc/leftPanel";
 import { SmallCenterPlate } from "../../components/hoc/plates/centerPlate";
 import { Header } from "../../components/ui/meny-time use/header";
 import { Modal } from "../../components/ui/meny-time use/modal";
 import { SomeContext } from "../../context";
 import { MainLoader } from "../../components/ui/meny-time use/loader";
-import { MeetingChild } from "../../childrens/meeting";
+import { MeetingChild } from "../../childrens/other/meeting";
 import { DftRPanel } from "../../components/hoc/dftrPanel";
 import ChangeTitle from "../../functions/ChangeTitle";
 import { Center } from "../../components/hoc/center";
 import { arrey } from "../../functions/GiveConst";
 import useDinamicPagination from "../../customHooks/useDinamicPagination";
+import { useParams } from "react-router";
 
 type Props = {
+
 }
 export const Meetings = ({ }: Props) => {
     const { loading, modal } = useContext<any>(SomeContext)
     const [meetings, setMeetings] = useState<any[]>([{}, {}, {}, {}, {}, {}, {}, {}])
+    const direction = useParams()
 
     const ref = useDinamicPagination(() => setMeetings((prev: any) => [...prev, ...arrey]))
 
@@ -39,7 +42,7 @@ export const Meetings = ({ }: Props) => {
 
                         <div ref={ref} className="scrollhandlerref"></div>
                     </Center>
-                    <DftRPanel direction={4} />
+                    <DftRPanel direction={Number(direction.iddirection)} />
                 </>
             </div>
         </>
