+++ src/pages/direction/Protokols.tsx
@@ -1,7 +1,6 @@
 import { useContext } from "react";
-import { ModalDirectionChildren } from "../../childrens/modalDirection";
+import { ModalDirectionChildren } from "../../childrens/other/modalDirection";
 import { LeftPanel } from "../../components/hoc/leftPanel";
-import { SmallCenterPlate } from "../../components/hoc/plates/centerPlate";
 import { Header } from "../../components/ui/meny-time use/header";
 import { Modal } from "../../components/ui/meny-time use/modal";
 import { SomeContext } from "../../context";
@@ -9,12 +8,15 @@ import { MainLoader } from "../../components/ui/meny-time use/loader";
 import { DftRPanel } from "../../components/hoc/dftrPanel";
 import ChangeTitle from "../../functions/ChangeTitle";
 import { Center } from "../../components/hoc/center";
+import { ProtokolsChild } from "../../childrens/pages/direction/protokols";
+import { useParams } from "react-router";
 
 type Props = {
 }
 export const Protokols = ({ }: Props) => {
     const { loading, modal } = useContext<any>(SomeContext)
     ChangeTitle('протоколы')
+    const direction = useParams()
 
     return (
         <>
@@ -26,19 +28,8 @@ export const Protokols = ({ }: Props) => {
                 }
                 <>
                     <LeftPanel function_={modal.SwapFn} />
-                    <Center>
-                        <SmallCenterPlate>
-                            <div className="dftcontainer">
-                                <span style={{ padding: '0 30px' }}>
-                                    <p style={{ margin: '10px 0 30px 0' }}>протоколы проведения встреч и организации турниров</p>
-                                    <a href="/svg/word.svg" download={''}>
-                                        <img src="/svg/word.svg" alt="" />
-                                    </a>
-                                </span>
-                            </div>
-                        </SmallCenterPlate>
-                    </Center>
-                    <DftRPanel direction={4} />
+                    <Center><ProtokolsChild /></Center>
+                    <DftRPanel direction={Number(direction.iddirection)} />
                 </>
             </div>
         </>
