+++ src/pages/direction/RegTeam.tsx
@@ -1,9 +1,7 @@
 import { useContext } from "react";
-import { ModalDirectionChildren } from "../../childrens/modalDirection";
+import { ModalDirectionChildren } from "../../childrens/other/modalDirection";
 import { LeftPanel } from "../../components/hoc/leftPanel";
-import { SmallCenterPlate } from "../../components/hoc/plates/centerPlate";
 import { RightPanel } from "../../components/hoc/rightPanel";
-import { InputFile, InputText } from "../../components/ui/meny-time use/customInput";
 import { Header } from "../../components/ui/meny-time use/header";
 import { Modal } from "../../components/ui/meny-time use/modal";
 import { SomeContext } from "../../context";
@@ -11,7 +9,7 @@ import { MainLoader } from "../../components/ui/meny-time use/loader";
 import ChangeTitle from "../../functions/ChangeTitle";
 import { Right } from "../../components/hoc/right";
 import { Center } from "../../components/hoc/center";
-import { Button } from "../../components/ui/meny-time use/customButton";
+import { RegTeamChild } from "../../childrens/pages/direction/regTeam";
 
 type Props = {
 
@@ -29,33 +27,8 @@ export const RegTeam = ({ }: Props) => {
             }
             <div className="main">
                 <LeftPanel function_={modal.SwapFn} />
-                <Center>
-                    <SmallCenterPlate>
-                        <div className="dftcontainer" style={{ minHeight: '500px', justifyContent: 'start', padding: '40px 0' }}>
-                            <form className="edit">
-                                <div style={{ width: '50%' }}>
-                                    <InputText title={"название"} value={""} setValue={() => undefined} max={0} />
-                                </div>
-
-                                <div style={{ width: '80%' }}>
-                                    <InputText title={"статус"} value={""} setValue={() => undefined} max={0} />
-                                </div>
-
-                                <span>
-                                    <InputFile setValue={() => undefined} title="фон команды" />
-                                    <InputFile setValue={() => undefined} title="изображение команды" />
-                                </span>
-
-                                <div style={{ display: 'flex', justifyContent: 'end', padding: '0 40px 0 0', margin: '20px 0 0 0' }}>
-                                    <Button title="сохранить" function_={() => undefined} />
-                                </div>
-                            </form>
-                        </div>
-                    </SmallCenterPlate>
-                </Center>
-                <Right>
-                    <RightPanel><div className="dftcontainer"></div></RightPanel>
-                </Right>
+                <Center><RegTeamChild /></Center>
+                <Right><RightPanel><div className="dftcontainer"></div></RightPanel></Right>
             </div>
         </>
     );
