+++ src/pages/direction/Tournaments.tsx
@@ -1,17 +1,18 @@
 import { useContext, useState } from "react";
-import { ModalDirectionChildren } from "../../childrens/modalDirection";
+import { ModalDirectionChildren } from "../../childrens/other/modalDirection";
 import { LeftPanel } from "../../components/hoc/leftPanel";
 import { SmallCenterPlate } from "../../components/hoc/plates/centerPlate";
 import { Header } from "../../components/ui/meny-time use/header";
 import { Modal } from "../../components/ui/meny-time use/modal";
 import { SomeContext } from "../../context";
 import { MainLoader } from "../../components/ui/meny-time use/loader";
-import { TournamentChild } from "../../childrens/tournament";
+import { TournamentChild } from "../../childrens/other/tournament";
 import { DftRPanel } from "../../components/hoc/dftrPanel";
 import ChangeTitle from "../../functions/ChangeTitle";
 import { Center } from "../../components/hoc/center";
 import { arrey } from "../../functions/GiveConst";
 import useDinamicPagination from "../../customHooks/useDinamicPagination";
+import { useParams } from "react-router";
 
 type Props = {
 }
@@ -20,7 +21,8 @@ export const Tournaments = ({ }: Props) => {
     const [tournameents, setTournaments] = useState<any[]>([{}, {}, {}, {}, {}, {}, {}, {}])
 
     const ref = useDinamicPagination(() => setTournaments((prev: any) => [...prev, ...arrey]))
-    
+    const direction = useParams()
+
     ChangeTitle('турниры')
     return (
         <>
@@ -39,8 +41,9 @@ export const Tournaments = ({ }: Props) => {
                         ))}
 
                         <div ref={ref} className="scrollhandlerref"></div>
+
                     </Center>
-                    <DftRPanel direction={4} />
+                    <DftRPanel direction={Number(direction.iddirection)} />
                 </>
             </div>
         </>
