+++ src/pages/main/1test.tsx
@@ -4,7 +4,7 @@ import { MainLoader } from "../../components/ui/meny-time use/loader";
 import { SomeContext } from "../../context";
 import { RightPanel } from "../../components/hoc/rightPanel";
 import { Modal } from "../../components/ui/meny-time use/modal";
-import { ModalDirectionChildren } from "../../childrens/modalDirection";
+import { ModalDirectionChildren } from "../../childrens/other/modalDirection";
 import { FullPlate } from "../../components/hoc/plates/fullPlate";
 import { LeftPanel } from "../../components/hoc/leftPanel";
 import ChangeTitle from "../../functions/ChangeTitle";
@@ -16,12 +16,12 @@ type Props = {
 }
 export const Test = ({ }: Props) => {
     const { loading, modal } = useContext<any>(SomeContext)
-    const plus = true
-    ChangeTitle('расходы')
+    ChangeTitle('ТЕСТ ПАДЕ')
     return (
         <>
             {modal.boolean && <Modal function_={modal.SwapFn}><ModalDirectionChildren function_={modal.SwapFn} /></Modal>}
             <Header />
+            <p>ТЕСТ ПАДЕ</p>
             <div className="main">
                 {loading &&
                     <MainLoader />
@@ -31,48 +31,11 @@ export const Test = ({ }: Props) => {
                     <Center>
 
                         <FullPlate>
-                            <div style={{ padding: '100px 0 0 0' }}>
-                                <div className="cashheader">
-                                    <div>вс 500 ₽</div>
-                                    <div>история расхода</div>
-                                    <div>дата <img src="/svg/calendar.svg" /></div>
-                                    <div>дисциплина</div>
-                                </div>
-                                <hr color="#262626" />
-                                <div className="cashheader greencash" style={{ padding: '10px 30px' }} id={'' && 'redcash'}>
-                                    <div>+5000 ₽</div>
-                                    <div>покупка мяча</div>
-                                    <div>20.05.06</div>
-                                    <div>dota</div>
-                                </div>
-                                <hr color="#262626" />
-                                <div className="cashheader greencash" style={{ padding: '10px 30px' }} id={plus && 'redcash'}>
-                                    <div>-5000 ₽</div>
-                                    <div>покупка мяча</div>
-                                    <div>20.05.06</div>
-                                    <div>dota</div>
-                                </div>
-                                <hr color="#262626" />
-                                <div className="cashheader greencash" style={{ padding: '10px 30px' }} id={plus && 'redcash'}>
-                                    <div>-5000 ₽</div>
-                                    <div>покупка мяча</div>
-                                    <div>20.05.06</div>
-                                    <div>dota</div>
-                                </div>
-                                <hr color="#262626" />
-                                <div className="cashheader greencash" style={{ padding: '10px 30px' }} id={'' && 'redcash'}>
-                                    <div>+5000 ₽</div>
-                                    <div>покупка мяча</div>
-                                    <div>20.05.06</div>
-                                    <div>dota</div>
-                                </div>
-                                <hr color="#262626" />
-                            </div>
+                            <>
+                            </>
                         </FullPlate>
                     </Center>
-                    <Right>
-                        <RightPanel><div className="dftcontainer"></div></RightPanel>
-                    </Right>
+                    <Right><RightPanel><div className="dftcontainer"></div></RightPanel></Right>
                 </>
             </div>
         </>
