+++ src/style/custom/input.scss
@@ -169,3 +169,15 @@ input[type=checkbox] {
 .smallinput>div {
     width: 35%;
 }
+
+select {
+    background-color: $background-color-dark;
+    padding: 10px;
+    font-size: 1em;
+    border-radius: 5px;
+    border: 1px solid $border-color;
+    outline: none;
+    cursor: pointer;
+    width: 100%;
+    min-height: 22px;
+}
