+++ src/style/custom/modal.scss
@@ -40,7 +40,9 @@
 .shadow {
     width: 100%;
     height: 100%;
-    background-color: rgba(0, 0, 0, 0.548);
+    background: #242424da;
+    backdrop-filter: blur(2px);
+    -webkit-backdrop-filter: blur(2px);
     position: fixed;
     z-index: 4;
     min-width: 199vh;
